break down auth
