+++ web-s3/src/context/useAuth.js
@@ -1,0 +1,5 @@
+import { useContext } from "react";
+import { AuthContext } from "./AuthContext";
+
+const useAuth = () => useContext(AuthContext);
+export default useAuth;
